+++ packages/lib/box-forge/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.70](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-box-forge@0.0.69...@opendaw/lib-box-forge@0.0.70) (2026-01-20)
+
+**Note:** Version bump only for package @opendaw/lib-box-forge
+
 ## [0.0.69](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-box-forge@0.0.68...@opendaw/lib-box-forge@0.0.69) (2026-01-16)
 
 **Note:** Version bump only for package @opendaw/lib-box-forge