+++ packages/lib/dawproject/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.53](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-dawproject@0.0.52...@opendaw/lib-dawproject@0.0.53) (2026-01-20)
+
+**Note:** Version bump only for package @opendaw/lib-dawproject
+
 ## [0.0.52](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-dawproject@0.0.51...@opendaw/lib-dawproject@0.0.52) (2026-01-16)
 
 **Note:** Version bump only for package @opendaw/lib-dawproject