+++ packages/lib/dom/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.69](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-dom@0.0.68...@opendaw/lib-dom@0.0.69) (2026-01-20)
+
+**Note:** Version bump only for package @opendaw/lib-dom
+
 ## [0.0.68](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-dom@0.0.67...@opendaw/lib-dom@0.0.68) (2026-01-16)
 
 **Note:** Version bump only for package @opendaw/lib-dom