+++ packages/studio/core-workers/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.73](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-core-workers@0.0.72...@opendaw/studio-core-workers@0.0.73) (2026-01-20)
+
+**Note:** Version bump only for package @opendaw/studio-core-workers
+
 ## [0.0.72](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-core-workers@0.0.71...@opendaw/studio-core-workers@0.0.72) (2026-01-16)
 
 **Note:** Version bump only for package @opendaw/studio-core-workers